// src/types/Tproject.ts
type Tproject = {
  id: number;
  email: string;
  researcher: string;
  piproject: string;
  acronym: string;
  nameProject: string;
  projectID: string;
  start: string;
  end: string;
  summary: string;
  websiteProject: string;
  fundingAgency: string;
  socialNetwork_x: string;
  socialNetwork_bsky: string;
  socialNetwork_inst: string;
  imageID_logo: string;
  imageID_FunAgen: string;
  imageID_ex1: string;
  imageID_ex2: string;
  imageID_ex3: string;
};

export default Tproject